requirements.txt
absl-py==2.3.0
asgiref==3.8.1
astunparse==1.6.3
boto3==1.38.36
botocore==1.38.36
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2.1
django-crum==0.7.9
django-rest-framework==0.1.0
django-storages==1.14.6
djangorestframework==3.16.0
flatbuffers==25.2.10
fonttools==4.58.1
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.13.0
idna==3.10
jmespath==1.0.1
keras==3.10.0
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
zopfli==0.2.3.post1
